gate access and list page updated

+++ src/app/modules/pages/gates/gates-list/gates-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { NgxUiLoaderService } from 'ngx-ui-loader';
 import { GateService } from 'src/app/modules/services/gateService/gate.service';
 import { ToastrService } from 'ngx-toastr';
+import { NgxCoolDialogsService } from 'ngx-cool-dialogs';
 declare var $;
 @Component({
   selector: 'app-gates-list',
@@ -10,18 +11,18 @@ declare var $;
 })
 export class GatesListComponent implements OnInit {
   public gatesList: any;
-  constructor(private gateService : GateService, 
+  constructor(private gateService : GateService, private coolDialogs: NgxCoolDialogsService,
     private loader: NgxUiLoaderService,
     private msgBox: ToastrService) { }
 
   ngOnInit(): void {
-
+   
     this.loader.start();
     this.gateService.getListOfGates().subscribe(data => {
       this.gatesList = data;
       this.loader.stop();
     }, error => {
-      this.msgBox.error(error.message);
+      this.msgBox.error(error.message,)
       this.loader.stop();
       
     }, () => {
@@ -37,6 +38,12 @@ export class GatesListComponent implements OnInit {
   }
 
   deleteGate(gateid: string){
+
+// Confirm
+this.coolDialogs.confirm('Are you sure you want to delete this gate ?')
+  .subscribe(res => {
+    if (res) {
+     
     this.loader.start();
     this.gateService.deleteGate(gateid).subscribe(data=>{      
       this.msgBox.info("Gate Deleted successfully.");
@@ -49,9 +56,12 @@ export class GatesListComponent implements OnInit {
       });
       
     },e=>{
-      this.msgBox.error("Something went wrong!! Unable to delete gate.");
+      this.msgBox.error("Something went wrong while deleting gate.");
       this.loader.stop();
     })  
+
+  } 
+});
   }
 
   openGateEditComponent(id){
@@ -66,6 +76,10 @@ export class GatesListComponent implements OnInit {
     }, () => {
       $('#tblGateList').DataTable().destroy();
       setTimeout(() => {
+
+
+
+        
         $('#tblGateList').DataTable({
           responsive: true,
           autoWidth: true,
